Add client from GUI;Do not add to pending clients with orders ""; Fixes the JSONs (but ask professor)

+++ src/Dish.java
@@ -3,9 +3,9 @@
 */
 
 public class Dish extends Product{
-    int Size;
+    float Size;
     
-    Dish(String newName, int newPrice, int newSize){
+    Dish(String newName, int newPrice, float newSize){
         super (newName, newPrice, 5);
         Size = newSize;
     }
